various compile fixes and an enhancement from znek

+++ src/luaxcore.cpp
@@ -1914,8 +1914,8 @@ static int lua_http_notify(lua_State* L)
     }
 
     fprintf(fp,
-        "NOTIFY %s HTTP/1.1\r\nHost: %s\r\nUser-Agent: %s\r\nConnection: close\r\nContent-Type: text/xml\r\nContent-Length: %i\r\n"
-        "NT: upnp:event\r\nNTS: upnp:propchange\r\nSID: uuid:%s\r\nSEQ: %i\r\nCache-Control: no-cache\r\n\r\n",url.urn,url.vhost,core::user_agent,(int)len,sid,seq);
+        "NOTIFY %s HTTP/1.1\r\nHost: %s\r\nUser-Agent: %s\r\nConnection: close\r\nContent-Type: text/xml\r\nContent-Length: %lu\r\n"
+        "NT: upnp:event\r\nNTS: upnp:propchange\r\nSID: uuid:%s\r\nSEQ: %i\r\nCache-Control: no-cache\r\n\r\n",url.urn,url.vhost,core::user_agent,(unsigned long)len,sid,seq);
     fwrite(data,len,1,fp);
     fflush(fp);
 
@@ -1974,7 +1974,7 @@ static int lua_http_download(lua_State* L)
                         url.urn,url.vhost,core::user_agent);
 
                 if(*post_data)
-                    fprintf(fp,"Content-Length: %d\r\n",(int)post_data_size);
+                    fprintf(fp,"Content-Length: %lu\r\n",(unsigned long)post_data_size);
 
 //                fprintf(fp,"Accept-Charset: utf-8\r\n");
